requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
bcrypt==4.0.1
nltk==3.8.1
spacy==3.6.1
pdfminer.six==20221105
python-docx==0.8.11
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.2
gunicorn==20.1.0
